pyproject.toml
[project]
name = "ai-website-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.0",
    "fastapi[standard]>=0.128.0",
    "gunicorn>=23.0.0",
    "livekit-agents[cartesia,openai,silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "openai>=2.15.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pip-system-certs>=5.3",
    "python-dotenv>=1.2.1",
    "langchain-openai",
    "langchain-chroma>=1.1.0",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1"
]
